requirements.txt
typer[all]
pyyaml
rich
